For #9605: move TestApplication to the helpers package.
The helpers/ package is created to separate the test support code from
the code under test.
