+++ app/src/test/java/org/mozilla/fenix/helpers/FenixRobolectricTestRunner.kt
@@ -4,7 +4,6 @@
 
 package org.mozilla.fenix.helpers
 
-import org.mozilla.fenix.TestApplication
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
