+++ app/src/test/java/org/mozilla/fenix/helpers/TestApplication.kt
@@ -2,8 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix
+package org.mozilla.fenix.helpers
 
+import org.mozilla.fenix.FenixApplication
 import org.mozilla.fenix.components.TestComponents
 
 class TestApplication : FenixApplication() {
